Cortex-hol-artifacts/Lab 4 - Agents/requirements.txt
snowflake-connector-python>=3.7.0
pyyaml>=6.0.1
